backend/pyproject.toml
[project]
name = "advisor-backend"
version = "0.1.0"
description = "Firehorse Advisor backend service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "google-cloud-firestore>=2.14.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "langgraph>=0.0.15",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.30",
    "yfinance>=0.2.36",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0",
    # Time series forecasting
    "statsmodels>=0.14.0",
    "scipy>=1.11.0",
    # Technical analysis
    "pandas-ta>=0.3.14b0",
    # Macro economic data
    "fredapi>=0.4.0",
    "openai>=2.16.0",
    "google-genai>=1.60.0",
    "langchain-google-genai>=4.2.0",
    "langchain>=1.2.7",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
# Ignore all warnings (e.g. from site-packages) except for our own code
filterwarnings = [
    "ignore",
    "ignore:.*mode.copy_on_write.*:UserWarning", # pandas_ta issue with pandas 3.0+
    "ignore:.*crypt.*:DeprecationWarning",       # passlib issue with python 3.13 slated removal
    "default:::app",
    "default:::tests",
]
